refactor(update): 拆分下载与响应解析模块
> This message is generated by AI.

`update/mod.rs` 同时承载下载 command、HTTP 客户端策略、流式文件写入、进度上报、取消处理、SHA-256 校验和响应文件名解析。后续引入 `UpdateManager` 时，这种职责集中会让状态迁移和下载行为混在同一个评审范围里。

因此，本 PR 先整理下载模块之间的接口，保持现有下载协议和运行行为：

```diff
 src-tauri/src/update/
-└── mod.rs       # 模块出口、下载流程和响应解析
+├── download.rs  # 下载控制流与 I/O 生命周期
+├── response.rs  # 响应文件名解析
+└── mod.rs       # 保留 Tauri command 出口
```

- `download.rs` 负责下载控制流、HTTP 客户端、会话取消、临时文件、进度上报和 SHA-256 校验。
- `response.rs` 负责 `Content-Disposition`、重定向 URL 和文件名清理。
- `update/mod.rs` 继续作为模块出口，保留 Tauri command 的现有注册路径。
- `DownloadProgressEvent` 和 `DownloadResult` 继续从 `crate::update` 导出。
- 现有响应解析测试随实现移到 `response.rs`。

响应解析是无副作用且可独立验证的逻辑，因此拥有单独的模块。HTTP client 仍与下载流程放在一起，因为代理、重定向和 User-Agent 策略都直接服务于该生命周期，单独的薄封装不会形成有价值的接口。

下载的 command 名称、参数、事件、camelCase 字段、session ID、旧任务取消、进度采样、临时文件清理、文件重命名和 SHA-256 校验保持不变。

本 PR 不包含 `UpdateManager`、Tauri managed state、安装模块、`changes.json`、增量更新路径处理、前端状态机和新的 HTTP client 抽象。
